+++ test/gtest/plugins/memory_handler.h
@@ -17,50 +17,62 @@
 #ifndef __MEMORY_HANDLER_H
 #define __MEMORY_HANDLER_H
 
+#include <absl/strings/str_format.h>
 #include "backend/backend_aux.h"
 #include "backend_engine.h"
 #include "common/nixl_log.h"
 #include "nixl.h"
 
-template<nixl_mem_t memType> class memoryHandler {
+template<nixl_mem_t memType> class memoryHandler;
+
+template<>
+class memoryHandler<DRAM_SEG> {
 public:
-    memoryHandler(size_t len, int dev_id) {
-        CHECK(false) << "memoryHandler() is not implemented for <" << memType << "> memory type";
+    memoryHandler(size_t len, int dev_id) : len_(len), dev_id_(dev_id) {
+        addr_ = malloc(len);
     }
 
     ~memoryHandler() {
-        CHECK(false) << "~memoryHandler() is not implemented for <" << memType << "> memory type";
+        free(addr_);
     }
 
     void
     set(char byte) {
-        CHECK(false) << "set() is not implemented for <" << memType << "> memory type";
+        for (size_t i = 0; i < len_; i++)
+            ((char *)addr_)[i] = byte + i;
     }
 
     bool
     check(char byte) {
-        CHECK(false) << "check() is not implemented for <" << memType << "> memory type";
-        return false;
+        for (size_t i = 0; i < len_; i++) {
+            uint8_t expected_byte = (uint8_t)byte + i;
+            if (((char *)addr_)[i] != expected_byte) {
+                NIXL_ERROR << "Verification failed at index " << i << "! local: " << ((char *)addr_)[i]
+                        << ", expected: " << expected_byte;
+                return false;
+            }
+        }
+        return true;
     }
 
     void
     reset() {
-        CHECK(false) << "reset() is not implemented for <" << memType << "> memory type";
+        memset(addr_, 0x00, len_);
     }
 
     void
     populateBlobDesc(nixlBlobDesc *desc, int buf_index = 0) {
-        CHECK(false) << "populateBlobDesc() is not implemented for <" << memType << "> memory type";
+        desc->addr = reinterpret_cast<uintptr_t>(addr_);
+        desc->len = len_;
+        desc->devId = dev_id_;
     }
 
     void
     populateMetaDesc(nixlMetaDesc *desc, int entry_index, size_t entry_size) {
-        CHECK(false) << "populateMetaDesc() is not implemented for <" << memType << "> memory type";
-    }
-
-    nixlBackendMD *
-    getMD() {
-        return md_;
+        desc->addr = reinterpret_cast<uintptr_t>(addr_) + entry_index * entry_size;
+        desc->len = entry_size;
+        desc->devId = dev_id_;
+        desc->metadataP = md_;
     }
 
     void
@@ -68,8 +80,69 @@ public:
         md_ = md;
     }
 
+    nixlBackendMD *
+    getMD() {
+        return md_;
+    }
+
 private:
     void *addr_;
+    size_t len_;
+    int dev_id_;
+    nixlBackendMD *md_;
+};
+
+template<>
+class memoryHandler<OBJ_SEG> {
+public:
+    memoryHandler(size_t len, int dev_id) : len_(len),
+                                            dev_id_(dev_id) {}
+
+    ~memoryHandler() = default;
+
+    void
+    set(char byte) {
+        CHECK(false) << "set() is not supported for OBJ_SEG type";
+    }
+
+    bool
+    check(char byte) {
+        CHECK(false) << "check() is not supported for OBJ_SEG type";
+        return false;
+    }
+
+    void
+    reset() {
+        CHECK(false) << "reset() is not supported for OBJ_SEG type";
+    }
+
+    void
+    populateBlobDesc(nixlBlobDesc *desc, int buf_index = 0) {
+        desc->addr = 0;
+        desc->len = len_;
+        desc->devId = dev_id_;
+        desc->metaInfo = absl::StrFormat("test-obj-key-%d", buf_index);
+    }
+
+    void
+    populateMetaDesc(nixlMetaDesc *desc, int entry_index, size_t entry_size) {
+        desc->addr = 0;
+        desc->len = len_;
+        desc->devId = dev_id_;
+        desc->metadataP = md_;
+    }
+
+    void
+    setMD(nixlBackendMD *md) {
+        md_ = md;
+    }
+
+    nixlBackendMD *
+    getMD() {
+        return md_;
+    }
+
+private:
     size_t len_;
     int dev_id_;
     nixlBackendMD *md_;
